Data: convert spellchecker format (single-doc)
Elasticsearch doesn't work well with the format coming from Isoxya
Pickax Spellchecker, even though it is a minimal representation of that
data structure. Change to project `results`, duplicating `paragraph`.

Note that this is single-doc only, which means although it improves the
matter somewhat, it does not go far enough (yet).