[#2276] Zone/Data,Resource: fix ES Bulk API errors not surfaced
Before the ES Bulk API was used, the response status was sufficient to
determine the status of the insert. This doesn't work for the Bulk API,
since some operations are allowed to fail, without affecting the
response status. Instead, `errors` is returned with the response body
indicating that one or more operations failed.

Fix by detecting `errors`, and throwing HTTP 400 if so. This is a little
simplistic, since in fact some operations might have succeeded, but the
deteriministic nature of `_id` usage should result in duplicates being
removed. More importantly, it fixes the issue where failures during
insert, for example because of syntax errors, are neither logged nor
retried.