[#2276] Zone/Data: fix ES auto-detected key types errors
ES throws errors when importing keys with types different to those already
auto-detected. Thus, we append the PlugProc tag as a suffix to the key. The
way of doing this here is hacky, and there is surely a better way -- perhaps
using lenses better, or maybe avoiding this whole approach and instead
handling it at the ToJSON instance level.